+++ lib/debouncer.dart
@@ -1,17 +1,22 @@
 import 'dart:async';
 import 'dart:ui';
 
-class Debouncer {
-  Debouncer._();
+class Debounce {
+  Debounce._();
 
   static const DEFAULT_DURATION = Duration(minutes: 1);
-  static final Debouncer instance = Debouncer._();
+  static final Debounce instance = Debounce._();
 
   Timer? _timer;
   void run(VoidCallback callback, {Duration duration = DEFAULT_DURATION}) {
-    _timer?.cancel();
-    _timer = Timer(duration, callback);
+    if (_timer == null || !(_timer?.isActive ?? false)) {
+      _timer = Timer(duration, () {
+        callback.call();
+        _timer?.cancel();
+        _timer = null;
+      },);
+    }
   }
 }
 
-final debounce = Debouncer.instance;
+final debounce = Debounce.instance;